requirements.txt
flet
requests
pyjnius